New code structure for the model's event API.
