+++ src-tauri/src/sketchbook/_model/_impl_session_state/_events_update_fns.rs
@@ -1,0 +1,6 @@
+use crate::sketchbook::ModelState;
+
+/// Implementation for events related to `update functions` of the model.
+impl ModelState {
+    // todo: add events to modify update function's expression
+}
